1/2 of the big 4
As 1/2 of the big four, Amazon, is a part of the big four company. This code includes the Amazon web scraping code; however, the code is not complete, and the code is also basic. This code deserves a potential of sophistication, furthermore, the code will be frequently updated with every chance it has.

Attachments: BeautifulSoup4 and Python3